Move callback to rendering module & rename
