+++ fabric-lifecycle-events-v1/src/client/resources/fabric-lifecycle-events-v1.client.mixins.json
@@ -3,7 +3,6 @@
   "package": "net.fabricmc.fabric.mixin.event.lifecycle.client",
   "compatibilityLevel": "JAVA_16",
   "client": [
-    "CameraMixin",
     "ClientChunkManagerMixin",
     "ClientPlayNetworkHandlerMixin",
     "ClientWorldClientEntityHandlerMixin",
