+++ fabric-lifecycle-events-v1/src/testmod/resources/fabric.mod.json
@@ -21,8 +21,7 @@
       "net.fabricmc.fabric.test.event.lifecycle.client.ClientBlockEntityLifecycleTests",
       "net.fabricmc.fabric.test.event.lifecycle.client.ClientEntityLifecycleTests",
       "net.fabricmc.fabric.test.event.lifecycle.client.ClientLifecycleTests",
-      "net.fabricmc.fabric.test.event.lifecycle.client.ClientTickTests",
-      "net.fabricmc.fabric.test.event.lifecycle.client.CameraUpdateTests"
+      "net.fabricmc.fabric.test.event.lifecycle.client.ClientTickTests"
     ]
   }
 }
