+++ fabric-rendering-v1/src/client/resources/fabric-rendering-v1.mixins.json
@@ -7,6 +7,7 @@
     "BlockColorsMixin",
     "BlockEntityRendererFactoriesMixin",
     "BuiltinModelItemRendererMixin",
+    "CameraMixin",
     "CapeFeatureRendererMixin",
     "DimensionEffectsAccessor",
     "EntityModelLayersAccessor",
