+++ fabric-rendering-v1/src/client/resources/fabric.mod.json
@@ -25,5 +25,6 @@
   ],
   "custom": {
     "fabric-api:module-lifecycle": "stable"
-  }
+  },
+  "accessWidener": "fabric-rendering-v1.accesswidener"
 }
