+++ fabric-rendering-v1/src/testmod/resources/fabric.mod.json
@@ -18,7 +18,8 @@
       "net.fabricmc.fabric.test.rendering.client.FeatureRendererTest",
       "net.fabricmc.fabric.test.rendering.client.TooltipComponentTests",
       "net.fabricmc.fabric.test.rendering.client.DimensionalRenderingTest",
-      "net.fabricmc.fabric.test.rendering.client.HudAndShaderTest"
+      "net.fabricmc.fabric.test.rendering.client.HudAndShaderTest",
+      "net.fabricmc.fabric.test.rendering.client.CameraUpdateTests"
     ]
   }
 }
